# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RealAddr for Agents API",
    "version": "0.8.0-spec",
    "description": "Implementation contract, not a deployed service. Hackathon mail scope is approval, enabled status and a human-entered destination; no physical forwarding. ENSv2 names on Ethereum Sepolia reference leases; public resolution does not disclose forwarding addresses."
  },
  "servers": [
    {
      "url": "https://address.chain.tokyo",
      "description": "Planned hackathon service origin. Domain and DNS setup are performed by the operator; deployment and availability are not yet verified."
    }
  ],
  "paths": {
...
    "/v1/auth/challenges": {
      "post": {
        "operationId": "createAgentChallenge",
        "summary": "Create domain-bound, expiring wallet challenge",
        "security": [],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Expiring wallet authentication challenge created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "400": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "401": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "403": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "404": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "409": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "410": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "422": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "429": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "503": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthChallengeRequest"
              }
            }
          }
        }
      }
    },
    "/v1/auth/sessions": {
      "post": {
        "operationId": "createAgentSession",
        "summary": "Consume challenge and issue agent credential",
        "security": [],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSession"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "400": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "401": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "403": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "404": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "409": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "410": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "422": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "429": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "503": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignatureProof"
              }
            }
          }
        }
      }
    },
    "/v1/auth/session": {
      "delete": {
        "operationId": "revokeAgentSession",
        "summary": "Revoke current agent credential",
        "security": [
          {
            "agentBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked",
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "400": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "401": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "403": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "404": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "409": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "410": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "422": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "429": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          },
          "503": {
            "description": "Request rejected. See error and traceId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Redacted diagnostic correlation ID"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "agentBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Application agent token, not World OAuth token."
      },
      "humanSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "realaddr_session",
        "description": "Server-side browser session. Each endpoint also enforces required wallet and World verification state."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound CSRF token; also verify Origin."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9]*(?:_[a-z0-9]+)*$",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "traceId": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "error",
          "message",
          "retryable",
          "traceId"
        ]
      },
...
      "AuthChallengeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "walletAddress": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{40}$"
          },
          "network": {
            "type": "string",
            "pattern": "^eip155:[0-9]+$"
          },
          "termsVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        },
        "required": [
          "walletAddress",
          "network",
          "termsVersion"
        ]
      },
      "Challenge": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "challengeId",
          "message",
          "expiresAt"
        ]
      },
...
    }
  }
}